+++ R/cv.coxtp.r
@@ -1,15 +1,16 @@
-#' fit a cross-validated Cox non-proportional hazards model with P-spline or Smoothing-spline where penalization tuning parameter is provided by cross validation
+#' fit a cross-validated Cox non-proportional hazards model with P-spline or Smoothing-spline where penalization tuning parameter is provided by cross-validation
 #' 
-#' Fit a Cox non-proportional hazards model via penalized maximum likelihood. 
+#' Fit a Cox non-proportional hazards model via penalized maximum likelihood. The penalization tuning parameter is provided by cross-validation.
 #' 
-#' @param event failure events response variable of length `nobs`, where `nobs` denotes the number of observations. It should be a vector containing 0 or 1.
+#' @param event failure event response variable of length `nobs`, where `nobs` denotes the number of observations. It should be a vector containing 0 or 1.
 #' @param z input covariate matrix, with `nobs` rows and `nvars` columns; each row is an observation. 
 #' @param time observed event time, which should be a vector with non-negative values.
 #' @param strata a vector of indicators for stratification. 
-#' Default = `NULL`, (i.e. no stratification group in the data), an unstratified model is implemented.
+#' Default = `NULL` (i.e. no stratification group in the data), an unstratified model is implemented.
 #' 
 #' @param penalty a character string specifying the spline term for the penalized Newton method. 
-#' This term is added to the log-partial likelihood, and the penalized log-partial likelihood serves as the new objective function to control the smoothness of the time-varying effects.
+#' This term is added to the log-partial likelihood, and the penalized log-partial likelihood serves as the new objective function to 
+#' control the smoothness of the time-varying coefficients.
 #' Default is `P-spline`. Three options are `P-spline`, `Smooth-spline` and `NULL`. 
 #' If `NULL`, the method will be the same as `coxtv` (unpenalized time-varying effects models) and `lambda` (defined below)
 #' will be set as 0. 
@@ -25,11 +26,11 @@
 #' 
 #' @param lambda a user specified sequence as the penalization coefficients in front of the spline term specified by `penalty`. 
 #' This is the tuning parameter for penalization.  The function `IC` can be used to select the best tuning parameter based on the information criteria. 
-#' Users can specify for larger values when the estimated time-varying effects are too high.
+#' Users can specify larger values when the absolute values of the estimated time-varying effects are too large.
 #' Default is `0` which refers to Newton method without penalization. 
 #' 
-#' @param nfolds number of folds for cross-validation - default is 5. The smallest value allowable is `nfolds`=3.
-#' @param foldid an optional vector of values between 1 and nfold identifying what fold each observation is in. If supplied, `nfolds` can be missing.
+#' @param nfolds number of folds for cross-validation, default is 5. The smallest value allowable is `nfolds`=3.
+#' @param foldid an optional vector of values between 1 and `nfolds` identifying what fold each observation is in. If supplied, `nfolds` can be missing.
 #'
 #' @param nsplines number of basis functions in the splines to span the time-varying effects, whose default value is 8. 
 #' We use the R function `splines::bs` to generate the B-splines. 
@@ -47,15 +48,18 @@
 #' If `penalty` is `Smooth-spline`, `degree`'s default value is 2. 
 #' 
 #' @param ties a character string specifying the method for tie handling. If there are no tied events, 
[…]
