+++ src/PenalizeStopCpp.cpp
@@ -1,5 +1,4 @@
 #define ARMA_NO_DEBUG
-#define ARMA_DONT_USE_OPENMP
 #define STRICT_R_HEADERS // needed on Windows, not macOS
 #include <RcppArmadillo.h>
 #include <omp.h>
