+++ src/PenalizeStopCpp.cpp
@@ -13,17 +13,6 @@ using namespace Rcpp;
 using namespace std;
 
 
-// // [[Rcpp::export]]
-// arma::uvec stra_sampling_cpp(int size, int prop){
-//   arma::uvec sample_id;
-  
-//   int s2      = floor(size/prop*1.0);
-//   sample_id   = arma::randperm(size,s2);
-
-//   return sample_id;
-// }
-
-
 List objfun_fixtra(const arma::mat &Z_tv, const arma::mat &B_spline, const arma::mat &theta,
                    const arma::mat &Z_ti, const arma::vec &beta_ti, const bool &ti,
                    const unsigned int n_strata,
@@ -865,7 +854,9 @@ List TIC_J_penalized_second_bresties(const arma::mat &Z_tv, const arma::mat &B_s
 }
 
 
-List spline_udpate(const arma::mat &Z_tv, const arma::mat &B_spline, arma::mat &theta, 
+List spline_udpate(const arma::mat &Z_tv, const arma::mat &B_spline, 
+                  const IntegerVector &count_strata,
+                  arma::mat &theta, 
                   const arma::mat &Z_ti, const arma::vec &beta_ti, 
                   arma::mat &S_matrix,
                   double &lambda_i,
@@ -1130,10 +1121,9 @@ List surtiver_fixtra_fit_penalizestop(const arma::vec &event, const IntegerVecto
   }
 
   arma::mat theta = theta_init; arma::vec beta_ti = beta_ti_init;
-  // double crit = 1.0, v = 1.0, logplkd_init = 0, logplkd, diff_logplkd, 
-  //   inc, rhs_btr = 0;
+
   double logplkd;
-  List objfun_list, update_list;
+  List objfun_list, update_list, hazard_list;
   NumericVector logplkd_vec, iter_NR_all;
   objfun_list = objfun_fixtra(Z_tv, B_spline, theta, Z_ti, beta_ti, ti, n_strata,
                               idx_B_sp, idx_fail, n_Z_strata, idx_Z_strata,
@@ -1163,7 +1153,6 @@ List surtiver_fixtra_fit_penalizestop(const arma::vec &event, const IntegerVecto
   arma::vec TIC2_secondterm = arma::zeros<arma::vec> (n_lambda);
   arma::vec GIC_secondterm = arma::zeros<arma::vec> (n_lambda);
   NumericVector AIC_all, TIC_all, TIC2_all, GIC_all;
-  //mat likelihood_mat = arma::zeros<arma::mat>(iter_max, n_lambda);    for fixed step usage
 
   arma::mat S_matrix;
   if(SplineType == "pspline") {
@@ -1175,15 +1164,13 @@ List surtiver_fixtra_fit_penalizestop(const arma::vec &event, const IntegerVecto
 
   arma::vec lambda_effct = arma::zeros<arma::vec> (p*K);
   arma::mat lambda_i_mat;
-  //List VarianceMatrix, VarianceMatrix2, VarianceMatrix_I, VarianceMatrix_J;
   arma::mat VarianceMatrix;
   arma::mat info;
 
   for (int i = 0; i < n_lambda; ++i)
   {
-    //generate a lambda matrix if different lambda is used for different covariate:
-    if(difflambda){
-
+    if(difflambda)
+    {
       for (int j = 0; j < p; ++j)
       {
         arma::vec sublambda = arma::zeros<arma::vec> (K);
@@ -1205,7 +1192,8 @@ List surtiver_fixtra_fit_penalizestop(const arma::vec &event, const IntegerVecto
       lambda_S_matrix = lambda_i_mat%S_matrix;
     }
 
-    SplineUdpate    = spline_udpate(Z_tv, B_spline, theta_ilambda,
[…]
