+++ src/RcppExports.cpp
@@ -149,42 +149,6 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// surtiver_fixtra_fit_penalizestop_lambdafromlarge
-List surtiver_fixtra_fit_penalizestop_lambdafromlarge(const arma::vec& event, const IntegerVector& count_strata, const arma::mat& Z_tv, const arma::mat& B_spline, const arma::mat& theta_init, const arma::mat& Z_ti, const arma::vec& beta_ti_init, const arma::vec& lambda_spline, const arma::mat& SmoothMatrix, const arma::vec& effectsize, const std::string& SplineType, const std::string& method, const double lambda, const double& factor, const bool& parallel, const unsigned int& threads, const double& tol, const unsigned int& iter_max, const double& s, const double& t, const std::string& btr, const std::string& stop, const bool& TIC_prox, const bool& fixedstep, const bool& difflambda, const bool& ICLastOnly);
-RcppExport SEXP _surtvep_surtiver_fixtra_fit_penalizestop_lambdafromlarge(SEXP eventSEXP, SEXP count_strataSEXP, SEXP Z_tvSEXP, SEXP B_splineSEXP, SEXP theta_initSEXP, SEXP Z_tiSEXP, SEXP beta_ti_initSEXP, SEXP lambda_splineSEXP, SEXP SmoothMatrixSEXP, SEXP effectsizeSEXP, SEXP SplineTypeSEXP, SEXP methodSEXP, SEXP lambdaSEXP, SEXP factorSEXP, SEXP parallelSEXP, SEXP threadsSEXP, SEXP tolSEXP, SEXP iter_maxSEXP, SEXP sSEXP, SEXP tSEXP, SEXP btrSEXP, SEXP stopSEXP, SEXP TIC_proxSEXP, SEXP fixedstepSEXP, SEXP difflambdaSEXP, SEXP ICLastOnlySEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< const arma::vec& >::type event(eventSEXP);
-    Rcpp::traits::input_parameter< const IntegerVector& >::type count_strata(count_strataSEXP);
-    Rcpp::traits::input_parameter< const arma::mat& >::type Z_tv(Z_tvSEXP);
-    Rcpp::traits::input_parameter< const arma::mat& >::type B_spline(B_splineSEXP);
-    Rcpp::traits::input_parameter< const arma::mat& >::type theta_init(theta_initSEXP);
-    Rcpp::traits::input_parameter< const arma::mat& >::type Z_ti(Z_tiSEXP);
-    Rcpp::traits::input_parameter< const arma::vec& >::type beta_ti_init(beta_ti_initSEXP);
-    Rcpp::traits::input_parameter< const arma::vec& >::type lambda_spline(lambda_splineSEXP);
-    Rcpp::traits::input_parameter< const arma::mat& >::type SmoothMatrix(SmoothMatrixSEXP);
-    Rcpp::traits::input_parameter< const arma::vec& >::type effectsize(effectsizeSEXP);
-    Rcpp::traits::input_parameter< const std::string& >::type SplineType(SplineTypeSEXP);
-    Rcpp::traits::input_parameter< const std::string& >::type method(methodSEXP);
-    Rcpp::traits::input_parameter< const double >::type lambda(lambdaSEXP);
-    Rcpp::traits::input_parameter< const double& >::type factor(factorSEXP);
[…]
