+++ JOSS/paper.bib
@@ -1,0 +1,255 @@
+
+@article{gaia,
+    author = {{Gaia Collaboration}},
+    title = "{The Gaia mission}",
+    journal = {Astronomy and Astrophysics},
+    archivePrefix = "arXiv",
+    eprint = {1609.04153},
+    primaryClass = "astro-ph.IM",
+    keywords = {space vehicles: instruments, Galaxy: structure, astrometry, parallaxes, proper motions, telescopes},
+    year = 2016,
+    month = nov,
+    volume = 595,
+    doi = {10.1051/0004-6361/201629272},
+    url = {http://adsabs.harvard.edu/abs/2016A%26A...595A...1G},
+}
+
+@article{astropy,
+    author = {{Astropy Collaboration}},
+    title = "{Astropy: A community Python package for astronomy}",
+    journal = {Astronomy and Astrophysics},
+    archivePrefix = "arXiv",
+    eprint = {1307.6212},
+    primaryClass = "astro-ph.IM",
+    keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
+    year = 2013,
+    month = oct,
+    volume = 558,
+    doi = {10.1051/0004-6361/201322068},
+    url = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
+}
+
+@misc{fidgit,
+  author = {A. M. Smith and K. Thaney and M. Hahnel},
+  title = {Fidgit: An ungodly union of GitHub and Figshare},
+  year = {2020},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/arfon/fidgit}
+}
+
+
+@article{baulies2015time,
+  title={Time-varying effect and long-term survival analysis in breast cancer patients treated with neoadjuvant chemotherapy},
+  author={Baulies, S and Belin, L and Mallon, P and Senechal, C and Pierga, JY and Cottu, P and Sablin, MP and Sastre, X and Asselain, B and Rouzier, R and others},
+  journal={British journal of cancer},
+  volume={113},
+  number={1},
+  pages={30--36},
+  year={2015},
+  publisher={Nature Publishing Group}
+}
+
+
+@article{brouwer2020time,
+  title  = {Time-varying survival effects for squamous cell carcinomas at oropharyngeal and nonoropharyngeal head and neck sites in the {United States}, 1973-2015},
+  author = {Brouwer, Andrew F and He, Kevin and Chinn, Steven B and Mondul, Alison M and Chapman, Christina H and Ryser, Marc D and Banerjee, Mousumi and Eisenberg, Marisa C and Meza, Rafael and Taylor, Jeremy MG},
+  journal={Cancer},
+  volume={126},
+  number={23},
+  pages={5137--5146},
+  year={2020},
+  publisher={Wiley Online Library}
+}
+
+@article{zucker1990nonparametric,
+  title={Nonparametric survival analysis with time-dependent covariate effects: a penalized partial likelihood approach},
+  author={Zucker, David M and Karr, Alan F},
+  journal={The Annals of Statistics},
+  volume={18},
+  number={1},
+  pages={329--353},
+  year={1990},
+  publisher={Institute of Mathematical Statistics}
+}
+
+@article{hastie1993varying,
+  title={Varying-coefficient models},
+  author={Hastie, Trevor and Tibshirani, Robert},
+  journal={Journal of the Royal Statistical Society: Series B (Methodological)},
+  volume={55},
+  number={4},
+  pages={757--779},
+  year={1993},
+  publisher={Wiley Online Library}
+}
+
+@article{gray1992flexible,
+  title={Flexible methods for analyzing survival data using splines, with applications to breast cancer prognosis},
+  author={Gray, Robert J},
[…]
