+++ JOSS/paper.md
@@ -1,0 +1,143 @@
+---
+title: 'surtvep: An R package for estimating time-varying effects'
+tags:
+  - R
+  - survival
+  - time-varying effects
+  - penalization
+output: pdf_document
+authors:
+  - name: Lingfeng Luo
+    affiliation: 1
+  - name: Wenbo Wu
+    affiliation: 2
+  - name: Jeremy Taylor
+    affiliation: 1
+  - name: Jian Kang
+    affiliation: 1
+  - name: Kevin He
+    corresponding: true #
+    affiliation: 1
+bibliography: paper.bib
+csl: apa.csl
+aas-doi: "10.3847/xxxxx <- update this with the DOI from AAS once you know it."
+aas-journal: " "
+affiliations:
+  - name: Department of Biostatistics, School of Public Health, University of Michigan
+    index: 1
+  - name: Departments of Population Health and Medicine, New York University Grossman School of Medicine
+    index: 2
+---
+
+**Authors:**
+
+Lingfeng Luo^1^, Wenbo Wu^2^, Jeremy Taylor^1^, Jian Kang^1^, Michael Kleinsasser^1^, Kevin He^1^*
+
+**Affiliations:**
+
+^1^ Department of Biostatistics, School of Public Health, University of Michigan
+
+^2^ Departments of Population Health and Medicine, New York University Grossman School of Medicine
+
+*Corresponding Author
+
+# Summary
+
+The `surtvep` package is an open-source software designed for estimating time-varying effects in survival analysis using the Cox non-proportional hazards model in R. With the rapid increase in large-scale time-to-event data from national disease registries, detecting and accounting for time-varying effects in medical studies have become crucial. Current software solutions often face computational issues such as memory limitations when handling large datasets. Furthermore, modeling time-varying effects for time-to-event data can be challenging due to small at-risk sets and numerical instability near the end of the follow-up period. `surtvep` addresses these challenges by implementing a computationally efficient Kronecker product-based proximal algorithm, supporting both unstratified and stratified models. The package also incorporates P-spline and smoothing spline penalties to improve estimation. Cross-validation and information criteria are available to determine the optimal tuning parameters. Parallel computation is enabled to further enhance computational efficiency. A variety of operating characteristics are provided, including estimated time-varying effects, confidence intervals, hypothesis testing, and estimated hazard functions and survival probabilities. The `surtvep` package thus offers a comprehensive and flexible solution to analyzing large-scale time-to-event data with dynamic effect trajectories.
+
+# Statement of Need
+
+The Cox non-proportional hazards model is a flexible and powerful tool for modeling time-varying effects of covariates in survival analysis. However, as the size of a dataset increases, the computational costs of this model can become substantial. Current software solutions, which may be effective for smaller datasets, face challenges when handling larger datasets. 
+
+Numerous studies have demonstrated the widespread presence of time-varying effects. For instance, the scientific literature has shown that factors like age, sex, and race can have non-constant associations with survival in cases such as end-stage renal disease [@he2017modeling; @he2022stratified], and breast cancer patients receiving neo-adjuvant chemotherapy and head and neck cancer patients [@baulies2015time; @brouwer2020time]. Ignoring the variations and relying solely on the Cox proportional hazards model can lead to inaccurate risk prediction and suboptimal treatment development.
+
[…]
