### 2023-07-23 #5644 @LingfengLuo0510
On Mon, Jul 10, 2023 at 7:08 AM The Open Journals editorial robot <
***@***.***> wrote:

> Hello human, I'm @editorialbot <https://github.com/editorialbot>, a robot
> that can help you with some common editorial tasks.
>
> For a list of things I can do to help you, just type:
>
> @editorialbot commands
>
> For example, to regenerate the paper pdf after making changes in the
> paper's md or bib files, type:
>
> @editorialbot generate pdf
>
