### 2023-07-23 #5644 @LingfengLuo0510
On Sun, Jul 23, 2023 at 4:48 PM The Open Journals editorial robot <
***@***.***> wrote:

> 👉📄 Download article proof
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.05644/joss.05644/10.21105.joss.05644.pdf>
> 📄 View article proof on GitHub
> <https://github.com/openjournals/joss-papers/blob/joss.05644/joss.05644/10.21105.joss.05644.pdf>
> 📄 👈
>
