### 2023-07-23 #5644 @LingfengLuo0510
Hi,

Thanks a lot! I have solved the issues and the paper can be compiled
successfully. Thank you so much! Looking forward to hearing from you!

Best,
Lingfeng

On Sun, Jul 23, 2023 at 1:27 PM Øystein Sørensen ***@***.***>
wrote:

> @LingfengLuo0510 <https://github.com/LingfengLuo0510> – please merge your
> default branch (with all of the source code) into your JOSS branch. Also,
> please rename your paper to be paper.md so @editorialbot
> <https://github.com/editorialbot> can detect the paper.
>
> Finally, we're managing a backlog of submissions right now and so please
> be patient while we wait for an editor to become available.
>
> @LingfengLuo0510 <https://github.com/LingfengLuo0510>, I'm the editor of
> this submission now. Please let me know when this issue is fixed, after
> which I can do an initial read of the paper and start looking for suitable
> reviewers.
>
