### 2023-07-24 #5644 @LingfengLuo0510
Hi,

WeakCha was a former team member of our research group. Therefore there is
a conflict of interest. May I ask if you can find other reviewers? Thank
you so much!

Best,
Lingfeng

On Mon, Jul 24, 2023 at 00:49 Øystein Sørensen ***@***.***>
wrote:

> 👋 @WeakCha <https://github.com/WeakCha> @turgeonmaxime
> <https://github.com/turgeonmaxime> @adibender
> <https://github.com/adibender>, would any of you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
