### 2023-07-24 #5688 @editorialbot
[REVIEW]: surtvep: An R package for estimating time-varying effects
**Submitting author:** @LingfengLuo0510 (<a href="http://orcid.org/0000-0002-7592-4748">Lingfeng Luo</a>)
**Repository:** https://github.com/UM-KevinHe/surtvep
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.0
