+++ R/coxtv.plot.R
@@ -51,7 +51,7 @@ plot.coxtv <- function(fit, parm, CI=TRUE, level=0.95, exponentiate=FALSE,
   defaultltys <- c("solid", "dashed", "dotted", "dotdash", "longdash")
   # if (missing(expand)) expand <- c(1,1)/100
   expand <- c(1,1)/100
-  ls.tvef <- confint(fit, times, parm, level)$tvef
+  ls.tvef <- confint(fit, time, parm, level)$tvef
   if (length(ls.tvef)==0) stop("No time-varying effect chosen!")
   # if (missing(labels)) labels <- names(ls.tvef)
   labels <- names(ls.tvef)
