+++ R/coxtp.base.R
@@ -550,7 +550,7 @@ VarianceMatrix <- function(formula, data, spline="P-spline", nsplines=8, ties="B
 #' @param fit fitted \code{"coxtp"} model.
 #' @param time the time interval to be estamtied. The default value is the time of the fitted model.
 #' @param parm the names of parameter.
-#' @param level the confidence level. Default is 0.95.
+#' @param level the confidence level. The default value is 0.95.
 #' 
 #' 
 #' @examples 
