### 2023-08-09 6415304
update notes and warnings
