+++ R/coxtp.base.R
@@ -1,4 +1,4 @@
-coxtp.base <- function(formula, data, spline="Smooth-spline", nsplines=8, ties="Breslow", 
+coxtp.base <- function(formula, data, spline="Smooth-spline", ties="Breslow", 
                      control, ...) {
   if (!ties%in%c("Breslow", "none")) stop("Invalid ties!")
   # pass ... args to coxtp.control
@@ -11,8 +11,10 @@ coxtp.base <- function(formula, data, spline="Smooth-spline", nsplines=8, ties="
                     names(extraArgs)[indx==0L]), domain=NA)
   }
   if (missing(control)) control <- coxtp.control(...)
-  degree <- control$degree
   
+  knots         <- control$knots
+  nsplines      <- control$nsplines
+  degree        <- control$degree
   TIC           <- control$TIC
   TIC_prox      <- control$TIC_prox
   penalize      <- control$penalize
@@ -21,7 +23,6 @@ coxtp.base <- function(formula, data, spline="Smooth-spline", nsplines=8, ties="
   fixedstep     <- control$fixedstep
   addsecond     <- control$addsecond
   ICLastOnly    <- control$ICLastOnly
-  
   
    # formula=fmla
   
@@ -67,10 +68,11 @@ coxtp.base <- function(formula, data, spline="Smooth-spline", nsplines=8, ties="
   nsplines <- nsplines[1]
   # model fitting
   if (spline=="P-spline") {
-    knots <- 
-      quantile(data[data[,term.event]==1,term.time], 
-               (1:(nsplines-degree-1))/(nsplines-degree))
-    
+    if(is.null(knots)){
+      knots <- 
+        quantile(data[data[,term.event]==1,term.time], 
+                 (1:(nsplines-degree-1))/(nsplines-degree))
+    }
     if (ties=="Breslow"){
       uniqfailtimes.str <- unname(unlist(lapply(
         split(data[data[,term.event]==1,term.time],
@@ -132,10 +134,12 @@ coxtp.base <- function(formula, data, spline="Smooth-spline", nsplines=8, ties="
       # return(fit)
     }
   } else if (spline=="Smooth-spline"){
-    if (ties=="Breslow"){
+    if(is.null(knots)){
       knots <- 
         quantile(data[data[,term.event]==1,term.time], 
                  (1:(nsplines-degree-1))/(nsplines-degree))
+    }
+    if (ties=="Breslow"){
       uniqfailtimes.str <- unname(unlist(lapply(
         split(data[data[,term.event]==1,term.time],
               data[data[,term.event]==1,term.str]), unique)))
@@ -177,6 +181,19 @@ coxtp.base <- function(formula, data, spline="Smooth-spline", nsplines=8, ties="
         }
       }
       SmoothMatrix  <- t(G_matrix)%*%W_matrix%*%G_matrix
+      print("start fitting:")
+      print(control$factor)
+      print(control$parallel)
+      print(control$threads)
+      print(control$iter.max)
+      print(control$tol)
+      print(control$s)
+      print(control$tau)
+      print(control$btr)
+      print(control$stop)
+      print(control$fixedstep)
+      print(control$gamma)
+      print(control$difflambda)
       fit <- 
         surtiver_fixtra_fit_penalizestop_bresties(data[,term.event], data[,term.time], 
                                                   count.strata, 
[…]
