### 2023-08-09 a987ba1
update check() notes
