### 2023-08-10 #5688 @turgeonmaxime
@LingfengLuo0510 I just read the manuscript for the first time, and I have a few questions/comments.

- I noticed that Figure 1 (the function flowchart) doesn't appear in the PDF. In its place, it looks like the caption is repeated twice. Pleas fix.
- In Figure 2, the confidence bands don't cover the same range as the curve estimate. Why is that?
- On line 119, you write: "We estimate cancer stage of kidney, lung, and breast." That's incorrect, you're estimating hazard ratios. The cancer stages are part of the data.
- For the section "Data Example", would it be possible to share your code?
- In the manuscript, could you clarify when a user should opt for a penalized approach? Does it depend on the number of covariates, or the number of basis functions?

I'm sure I'll have further comments/questions as I keep reviewing the package, but I'll put them in a different comment!
