### 2023-08-10 #5688 @turgeonmaxime
Also, I haven't been able to install the package using the instructions in the manuscript. Here is the error message I got.
```r
r$> remotes::install_github('UM-KevinHe/surtvep')
Downloading GitHub repo UM-KevinHe/surtvep@HEAD
── R CMD build ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
   checking for file ‘/private/var/folders/hb/r5l2r5b12gg5y2zxyg4l7nh80000gn/T/Rtmp42cwos/remotes1030ea884ab0/UM-KevinHe-surtvep-31f9fd4/DESC✔  checking for file ‘/private/var/folders/hb/r5l2r5b12gg5y2zxyg4l7nh80000gn/T/Rtmp42cwos/remotes1030ea884ab0/UM-KevinHe-surtvep-31f9fd4/DESCRIPTION’
─  preparing ‘surtvep’:
─  cleaning src
─  building ‘surtvep_1.0.0.tar.gz’

* installing *source* package ‘surtvep’ ...
** using staged installation
** libs
using C++ compiler: ‘Apple clang version 14.0.3 (clang-1403.0.22.14.1)’
using C++11
using SDK: ‘MacOSX13.3.sdk’
clang++ -arch arm64 -std=gnu++11 -I"/Library/Frameworks/R.framework/Resources/include" -DNDEBUG  -I'/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/Rcpp/include' -I'/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/RcppArmadillo/include' -I/opt/R/arm64/include     -fPIC  -falign-functions=64 -Wall -g -O2  -c PenalizeStopCpp.cpp -o PenalizeStopCpp.o
PenalizeStopCpp.cpp:4:10: fatal error: 'omp.h' file not found
#include <omp.h>
         ^~~~~~~
1 error generated.
make: *** [PenalizeStopCpp.o] Error 1
ERROR: compilation failed for package ‘surtvep’
* removing ‘/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/surtvep’
* restoring previous ‘/Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/library/surtvep’
Warning message:
In i.p(...) :
  installation of package ‘/var/folders/hb/r5l2r5b12gg5y2zxyg4l7nh80000gn/T//Rtmp42cwos/file1030e79df6e36/surtvep_1.0.0.tar.gz’ had non-zero exit status
```

I did notice that the instructions in the `README.md` file are slightly different, as they point to the branch `openmp`. When I follow those instructions (i.e. `remotes::install_github("UM-KevinHe/surtvep", ref = "openmp")`, then the installation proceeds without error.

For reference, here's my session information:

```r
r$> sessionInfo()
R version 4.3.0 (2023-04-21)
Platform: aarch64-apple-darwin20 (64-bit)
Running under: macOS Ventura 13.4.1

Matrix products: default
BLAS:   /Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/lib/libRblas.0.dylib
LAPACK: /Library/Frameworks/R.framework/Versions/4.3-arm64/Resources/lib/libRlapack.dylib;  LAPACK version 3.11.0

locale:
[1] en_CA.UTF-8/en_CA.UTF-8/en_CA.UTF-8/C/en_CA.UTF-8/en_CA.UTF-8

time zone: America/Edmonton
tzcode source: internal

attached base packages:

 [1] processx_3.8.2    compiler_4.3.0    R6_2.5.1          rprojroot_2.0.3   cli_3.6.1         prettyunits_1.1.1 tools_4.3.0
 [8] curl_5.0.1        crayon_1.5.2      remotes_2.4.2     desc_1.4.2        callr_3.7.3       ps_1.7.5          pkgbuild_1.4.2
```
