### 2023-08-10 #5688 @turgeonmaxime
@LingfengLuo0510 One more thing: I ran `devtools::check()` on the `openmp` branch, and it highlighted many issues with the documentation (among other things). I suggest you ran `devtools::check()` yourself and fix the warnings, and consider fixing the notes too.
