+++ R/coxtp.plot.R
@@ -19,6 +19,7 @@
 #' @param time the time points for which the time-varying coefficients to be plotted. 
 #' The default value is the unique observed event times in the dataset fitting the time-varying effects model.
 #' @param allinone if `TRUE`, the time-varying trajectories for different covariates are combined into a single plot. The default value is `FALSE`.
+#' @param \dots Other graphical parameters to plot
 #' 
 #' @importFrom ggplot2 ggplot aes geom_line geom_ribbon theme_bw theme element_text element_blank element_line margin labs ggtitle geom_hline scale_x_continuous scale_y_continuous scale_linetype_manual scale_color_manual guides guide_legend scale_fill_manual 
 #' @importFrom rlang .data
@@ -33,10 +34,9 @@
 #' event <- ExampleData$event
 #' fit <- coxtp(event = event, z = z, time = time)
 #' plot(fit$lambda1)
-#' 
 plot.coxtp <- function(x, parm, CI=TRUE, level=0.95, exponentiate=FALSE, 
                        xlab, ylab, xlim, ylim, allinone=FALSE, 
-                       title, linetype, color, fill, time) {
+                       title, linetype, color, fill, time, ...) {
   
   if (missing(x)) stop ("Argument x is required!")
   fit <- x
