+++ R/confint.coxtv.R
@@ -2,12 +2,12 @@
 #' 
 #' Get confidence intervals of time-varying coefficients from a fitted `coxtv` or `coxtp` object. 
 #' 
-#' @param fit fitted \code{"coxtv"} model.
+#' @param object fitted \code{"coxtv"} model.
 #' @param time the time points for which the confidence intervals to be estimated. 
 #' The default value is the unique observed event times in the dataset fitting the time-varying effects model.
 #' @param parm the names of parameters.
 #' @param level the confidence level. The default value is 0.95.
-#' 
+#' @param \dots other parameters to function
 #' 
 #' @examples 
 #' \dontrun{
@@ -20,10 +20,10 @@
 #' }
 #' 
 #' @exportS3Method confint coxtv
-confint.coxtv <- function(fit, time, parm, level=0.95) {
-  if (missing(fit)) stop ("Argument fit is required!")
-  if (!inherits(fit,"coxtv")) stop("Object fit is not of class 'coxtv'!")
-  
+confint.coxtv <- function(object, time, parm, level=0.95, ...) {
+  if (missing(object)) stop ("Argument object is required!")
+  if (!inherits(object,"coxtv")) stop("object is not of class 'coxtv'!")
+  fit <- object
   if (missing(time)) {
     time <- fit$times
   } else {
