### 2023-08-17 6b0ee4f
fix check notes support openmp
