+++ JOSS/GenerateRealDataPlot.R
@@ -1,0 +1,205 @@
+rm(list = ls())
+setwd("/Users/lingfengluo/Dropbox (University of Michigan)/Lingfeng Research/R Tutorial Package/surtvep/")
+
+library(devtools)
+
+########################################################################
+######fit based on kidney/lung/breast cancer data:
+########################################################################
+
+################################################################################################################################################
+#get time-to-event data from kidney (z_kidney, event_kidney, and time_kidney)
+model.kidney  <- coxtp(z = z_kidney, event = event_kidney, time=time_kidney, lambda = seq(0.1,1,0.1) * ((dim(z)[1])^(1/3)),
+                       threads = 3, parallel = TRUE)
+IC.kidney <- IC(model.kidney)
+
+
+beta_cancer_kidney_cancerall        <- get.tvcoef(IC.kidney$model.TIC)
+colnames(beta_cancer_kidney_cancerall) <- c("age_50","age_50_59","age_70", 
+                                            "race_black", "race_hispanic", "race_asian", "race_other", 
+                                            "yod_2009_2012","yod_2013_2017",
+                                            "distant", "regional_both", "regional_ext", "regional_lymph", "stage_unknown")
+beta_cancer_kidney_cancerall$time   <- time_kidney
+beta_cancer_kidney_cancerall$constant   <- 1
+
+confint_kidney <- confint(IC.kidney$model.TIC)
+beta_cancer_kidney_cancerall$distant_low <- confint_kidney$tvef$distant[,2]
+beta_cancer_kidney_cancerall$distant_up <- confint_kidney$tvef$distant[,3]
+beta_cancer_kidney_cancerall$regional_both_low <- confint_kidney$tvef$regional_both[,2]
+beta_cancer_kidney_cancerall$regional_both_up <- confint_kidney$tvef$regional_both[,3]
+beta_cancer_kidney_cancerall$regional_ext_low <- confint_kidney$tvef$regional_ext[,2]
+beta_cancer_kidney_cancerall$regional_ext_up <- confint_kidney$tvef$regional_ext[,3]
+beta_cancer_kidney_cancerall$regional_ext_low <- confint_kidney$tvef$regional_ext[,2]
+beta_cancer_kidney_cancerall$regional_ext_up <- confint_kidney$tvef$regional_ext[,3]
+beta_cancer_kidney_cancerall$regional_lymph_low <- confint_kidney$tvef$regional_lymph[,2]
+beta_cancer_kidney_cancerall$regional_lymph_up <- confint_kidney$tvef$regional_lymph[,3]
+################################################################################################################################################
+
+################################################################################################################################################
+#get time-to-event data from lung (z_lung, event_lung, and time_lung)
+model.lung  <- coxtp(z = z_lung, event = event_lung, time=time_lung, lambda = seq(0.1,1,0.1) * ((dim(z)[1])^(1/3)),
+                     threads = 3, parallel = TRUE)
+IC.lung <- IC(model.lung)
+
+
+beta_cancer_lung_cancerall        <- get.tvcoef(IC.lung$model.TIC)
+colnames(beta_cancer_lung_cancerall) <- c("age_50","age_50_59","age_70", 
+                                          "race_black", "race_hispanic", "race_asian", "race_other", 
[…]
