### 2023-08-24 #5688 @osorensen
@LingfengLuo0510, I just wanted to add to the comments from @turgeonmaxime that although we don't require R package submissions to be on CRAN, they should in general pass `R CMD check` with no NOTEs, WARNINGs, or ERRORs, unless there are very good arguments for not doing so. However, since this is an R package that needs compilation, I would also strongly recommend uploading it to CRAN, as it makes the installation process much easier for most users.
