### 2023-09-03 #5688 @LingfengLuo0510
@osorensen Thanks a lot for the instructions. I am submitting this project to CRAN now. I will update when it is successfully uploaded.
@turgeonmaxime The installation issues come from the fact that OpenMP feature is naturally supported in the Mac OS m1 chip system. I am uploading the package to CRAN and trying to fix the issue. I will reply to the points soon. Thanks a lot for the comments!
