+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: surtvep
-Title: Cox non-proportional hazards model with time-varying coefficients
+Title: Cox Non-Proportional Hazards Model with Time-Varying Coefficients
 Version: 1.0.0
 URL: https://github.com/UM-KevinHe/surtvep,
     https://um-kevinhe.github.io/surtvep/
@@ -9,10 +9,13 @@ Authors@R: c(
     person("Wenbo","Wu",role=c("aut"),email="wenbowu@umich.edu"),
     person("Kevin","He",role=c("aut"),email="kevinhe@umich.edu")
   )
-Description: surtvep is an R package for fitting Cox non-proportional hazards models with time-varying coefficients. 
-  Both unpenalized procedures (Newton and proximal Newton) and penalized procedures (P-splines and smoothing splines) are included using B-spline basis functions for estimating time-varying coefficients. 
-  For penalized procedures, cross validations, mAIC, TIC or GIC are implemented to select tuning parameters. 
-  Utilities for carrying out post-estimation visualization, summarization, point-wise confidence interval and hypothesis testing are also provided.
+Description: Fit Cox non-proportional hazards models with time-varying coefficients. 
+    Both unpenalized procedures (Newton and proximal Newton) and penalized procedures 
+    (P-splines and smoothing splines) are included using B-spline basis functions for 
+    estimating time-varying coefficients. For penalized procedures, cross validations, 
+    mAIC, TIC or GIC are implemented to select tuning parameters. Utilities for 
+    carrying out post-estimation visualization, summarization, point-wise confidence 
+    interval and hypothesis testing are also provided.
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
