### 2023-10-09 e62b4af
delete vignettes
