+++ vignettes/Model_Parameters.Rmd
@@ -1,168 +1,0 @@
----
-title: "Model Parameters"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Model Parameters}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-
-# 1. Functions
-
-## 1.1 `coxtp`:
-
-`coxtp(event,z,time,strata=c(),spline="P-spline", nsplines=8, ties="Breslow",tol=1e-9, iter.max=20L, method="Newton",`
-`lambda=1e8,factor=10, btr="dynamic", sigma=1e-2, tau=0.6,stop="incre", parallel=FALSE, threads=1L,`
-`degree=3L, TIC = FALSE, TIC_prox = FALSE, lambda_spline = 0, ord = 4, fixedstep = FALSE, effectsize = 0,`
-`difflambda = FALSE,addsecond = FALSE, penalizestop = FALSE, ICLastOnly = FALSE)`
-
-The main function in our package is `coxtp` which refers to the "Cox Non-proportional Hazards model with penalization". 
-
-
-
-### 1.1.1 Common Parameters
-
-`coxtp` function have two required input, which is `formula` and `data`.
-
-* **`event`**: Event vector, should be a vector containing 0 or 1
-* **`z`**: Covariate matrix
-* **`time`**: Time vector, should be a vector with non-negative numeric value 
-* **`strata`**: stratification group defined in the data. If there exist stratification group, please enter as vector.
-
-
-### 1.1.2 Method related Parameters:
-
-Here we offered both model estimation with penalization and without penalization. Detail method could be found in our paper(add link here). Following parameters defines the model settings:
-
-* **`spline`**: The spline term for Penalized Newton's Method(Add section Number related to the paper). Default setting is **`spline="Smooth-spline"`**
-  * **`"P-spline"`**: The P-splines are low rank smoother using a B-spline basis. Detail information about P-spline could be found in (Add section number and link)
-  * **`"Smooth-spline"`** : (Add section number and link)
-* **`nsplines`** : Number of base functions in the B-splines, default is 8.
-
-* **`method`** : Selecting Method used, default is **`method="Newton"`**
-  * **`"Newton"`**: Usual Newton’s Approach.
-  * **`"ProxN"`** : Proximal Newton–Raphson Method proposed in our paper(get link here).
-* **`lambda`** : Parameter for Proximal Newton's Method. Default is **`lambda=1e8`**
-* **`tau`** : (Alpha or beta?) in the Newton's Method, Default is **`tau=0.5`**. Used to control for step size.
-* **`lambda_spline`** :  Smoothing parameter lambda. Default is **`lambda_spline = 0`** which refers to Newton's Method without penalization.
-
-* **`degree`**: Degree of smoothing spline. Default setting is **`degree=3L`**.
-  * **`degree=2L`** refers to Smoothing quadratic spline.
-  * **`degree=3L`** refers to Smoothing cubic spline.
-  
-* **`ord`** : Specify which derivative to penalize. Default setting is **`ord=4`**.
-  * **`ord=4`**: Penalized second order derivative, suggest to use this when using the smoothing cubic spline.
-  * **`ord=3`**: Penalized first order derivative, suggest to use this when using the smoothing quadratic spline.
-  
-* `penalizestop`: could delete
-  
-## 1.1.3 Stopping creteria related Parameters:
-
-* **`tol`**: Convergence threshold. The default threshold is set as **`tol=1e-6`**
-* **`iter.max`**: Maximum Iteration number, default is **`iter.max=20L`**
[…]
