+++ vignettes/reference.Rmd
@@ -1,25 +1,0 @@
----
-title: "Reference"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Reference}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-
-
-# Related Paper:
-insert paper link here
-
-
-# Our main web
-
-insert web link here
