+++ _pkgdown.yml
@@ -1,64 +1,0 @@
-url: https://um-kevinhe.github.io/surtvep/
-template:
-  params:
-    bootswatch: lumen
-    ganalytics: UA-54921687-5
-    mathjax: true
-    
-vignettes:
-  - vignettes/surtvep.Rmd
-
-
-authors:
-  Kevin He:
-    href: "https://www.um-kevinhe-group.org/"
-  Wenbo Wu:
-    href: "https://www.wenbowu.me/"
-    
-navbar:
- components:
-   articles:
-    text: Appendix
-    menu:
-    - text: Category A
-    - text: B-spline construction
-      href: articles/Appendix.html
-    - text: Title A2
-      href: articles/a2.html
-    - text: -------
-    - text: "Category B"
-    - text: Plotting
-      href: articles/plotting.html
-      
-reference:
-- title: Main Fitting Functions
-  desc: The main function to fit the Cox-non proportional hazards model
-  contents:
-  - coxtv
-  - coxtp
-  - cv.coxtp
-- title: Information Criteria
-  desc: Calculating the Information Criteria to select the tunning paratmeter for
-    the penalized methods
-  contents: IC
-- title: Tests of Cox-non proportional hazards model
-  contents:
-  - tvef.ph
-  - tvef.zero
-  - tvef.zero.time
-- title: Helper functions
-  contents:
-  - get.tvcoef
-  - confint.coxtp
-  - confint.coxtv
-  - plot.coxtp
-  - plot.coxtv
-  - baseline
-  - plot.baseline
-- title: Package data sets
-  contents:
-  - ExampleData
-  - ExampleDataBinary
-  - StrataExample
-  - support
-

[…]
