### 2023-10-12 77937b7
reset options for user
