+++ R/coxtv.plot.R
@@ -38,7 +38,8 @@
 plot.coxtv <- function(x, parm, CI=TRUE, level=0.95, exponentiate=FALSE, 
                        xlab, ylab, xlim, ylim, allinone=FALSE, 
                        title, linetype, color, fill, time, ...) {
-  
+  old <- options()
+  on.exit(options(old))
   if (missing(x)) stop ("Argument x is required!")
   fit <- x
   if (!inherits(fit,"coxtv")) stop("Object fit is not of class 'coxtv'!")
