### 2023-10-12 #5688 @adibender
@LingfengLuo0510 some things that are currently missing from your package:
- LICENSE file
- automated tests (e.g. using testthat package)
- Community guidelines (how to contribute)
