+++ R/coxtv.plot.R
@@ -21,6 +21,8 @@
 #' @param allinone if `TRUE`, the time-varying trajectories for different covariates are combined into a single plot. The default value is `FALSE`.
 #' @param \dots other graphical parameters to plot
 #' 
+#' @return A plot is produced, and nothing is returned.
+#' 
 #' @importFrom ggplot2 ggplot aes geom_line geom_ribbon theme_bw theme element_text element_blank element_line margin labs ggtitle geom_hline scale_x_continuous scale_y_continuous scale_linetype_manual scale_fill_manual 
 #' @importFrom ggpubr annotate_figure ggarrange text_grob
 #' @importFrom rlang .data
