+++ man/support.Rd
@@ -87,7 +87,6 @@ covariates for activities of daily living. Therefore, surrogate activities
 of daily living were removed. See details in the R package (casebase) by Bhatnagar et al. (2020).
 }
 \examples{
-\dontrun{
 data(support)
 
 support <- support[support$ca \%in\% c("metastatic"),]
@@ -112,15 +111,8 @@ z_age <- model.matrix(~age)[,-1]
 z <- data.frame(z_age, sex, diabetes)
 colnames(z) <- c("age_50", "age_50_59", "age_70", "diabetes", "male")
 
-library(survminer)
-library(survival)
 data <- data.frame(time, death, z)
-fit1 <- survfit(Surv(time, death) ~ diabetes, data = data)
-fit2 <- survfit(Surv(time, death) ~ age_50 + age_50_59 + age_70, data = data)
-ggsurvplot(fit1, data = data)
-ggsurvplot(fit2, data = data, legend.labs = c("60-69", "70+", "50-59", "<50"))
 fit.coxtv <- coxtv(event = death, z = z, time = time)
-}
 }
 \references{
 Bhatnagar, S., Turgeon, M., Islam, J., Hanley, J. A., and Saarela, O. (2020) casebase: Fitting Flexible Smooth-in-Time
