+++ man/support.Rd
@@ -88,17 +88,12 @@ of daily living were removed. See details in the R package (casebase) by Bhatnag
 }
 \examples{
 data(support)
-
 support <- support[support$ca \%in\% c("metastatic"),]
-
 time <- support$d.time
 death <- support$death
-
 diabetes <-  model.matrix(~factor(support$diabetes))[,-1]
-
 #sex: female as the reference group
 sex <- model.matrix(~support$sex)[,-1]
-
 #age: continuous variable
 age <-support$age
 age[support$age<=50] <- "<50"
@@ -107,10 +102,8 @@ age[support$age>60 & support$age<70] <- "60-69"
 age[support$age>=70] <- "70+"
 age <- factor(age, levels = c("60-69", "<50", "50-59", "70+"))
 z_age <- model.matrix(~age)[,-1]
-
 z <- data.frame(z_age, sex, diabetes)
 colnames(z) <- c("age_50", "age_50_59", "age_70", "diabetes", "male")
-
 data <- data.frame(time, death, z)
 fit.coxtv <- coxtv(event = death, z = z, time = time)
 }
