+++ README.md
@@ -23,10 +23,13 @@ In our statistical software tutorial, we address a common issue encountered when
 
 The package only works for R 4.1.0+.
 
-You can install 'surtvep' via:
+You can install 'surtvep' via CRAN or github:
 
-    install.packages("devtools")
-    install.packages("remotes")
+    install.packages("surtvep")
+
+    #or
+    require("devtools")
+    require("remotes")
     remotes::install_github("UM-KevinHe/surtvep")
 
 We recommand to start with <a href="https://um-kevinhe.github.io/surtvep/articles/surtvep.html#quick-start" target="_blank">tutorial</a>, as it provides an overview of the package's usage, including preprocessing, model training, selection of penalization parameters, and post-estimation procedures.
