### 2023-11-17 #5688 @osorensen
@LingfengLuo0510, I notice the package has been on CRAN for a while now. Could you please let us know if there's also been any progress in addressing the points raised by the reviewers?
