+++ JOSS/paper.md
@@ -112,7 +112,7 @@ plot(fit.ic$mAIC, ylim = c(-3,10))
 ![The estimated time-varying coefficients (log hazard ratio) from `coxtv` and `coxtp`. The tuning parameter for `coxtp` is selected using mAIC.  \label{fig:coxtv}](coxtv_tp.png)
 
 
-In utilizing the `coxtv` and `coxtp` functions, users have the flexibility to choose based on their dataset's specifications. Numerical instabilities are commonly encountered when analyzing survival data of a small sample size or when the data includes some binary covariates with proportions that approach either zero or one. In these scenarios, the second-order information matrix can become ill-conditioned. See discussion in [@wu2022scalable; @lfluo2023smmr]. In these intricacies, we recommend the use of the penalized method, `coxtp`. When determining the number of basis functions, a typical range is between 5-10. Though the choice is somewhat flexible, it has limited impact on results unless set too small [@gray1992flexible]. Users might consider increasing this number when applying the penalized method.
+In utilizing the `coxtv` and `coxtp` functions, users have the flexibility to choose based on their dataset's specifications. Numerical instabilities are commonly encountered when analyzing survival data of a small sample size or when the data includes some binary covariates with proportions that approach either zero or one. In these scenarios, the second-order information matrix can become ill-conditioned. See discussion in [@wu2022scalable; @lfluo2023smmr]. To address this issue, the employment of the penalized method `coxtp` is recommended. When determining the number of basis functions, a typical range is between 5-10. Though the choice is somewhat flexible, it has limited impact on results unless set too small [@gray1992flexible]. Users might consider increasing this number when applying the penalized method.
 
 
 # Data Example 
