### 2023-11-25 #5688 @LingfengLuo0510
@turgeonmaxime

Thanks a lot for your comments.

1. "I noticed that Figure 1 (the function flowchart) doesn't appear in the PDF. In its place, it looks like the caption is repeated twice. Please fix."
Fixed. See the revised Figure 1.

2. "In Figure 2, the confidence bands don't cover the same range as the curve estimate. Why is that?"
The y-axis range in the updated Figure 2 has been expanded to ensure complete coverage of the confidence bands and the curve estimate. See the updates in Figure 2.

3. "On line 119, you write: "We estimate cancer stage of kidney, lung, and breast." That's incorrect, you're estimating hazard ratios. The cancer stages are part of the data."
See the revised text on line 128 for the correct information.

4. "For the section "Data Example", would it be possible to share your code?"
The code for model fitting and plotting can be found in the Repository-joss branch. The name of the file is "GenerateRealDataPlot.R". Access to data can be requested at https://seer.cancer.gov/data/access.html.

5. "In the manuscript, could you clarify when a user should opt for a penalized approach? Does it depend on the number of covariates, or the number of basis functions?"
We have added a detailed paragraph on line 16 to clarify when a user should opt for a penalized approach.

6. "Also, I haven't been able to install the package using the instructions in the manuscript. Here is the error message I got."
The package is now published on CRAN. See line 105.
