### 2023-12-04 #5688 @turgeonmaxime
@LingfengLuo0510 Thank you for addressing my previous comments. I especially appreciate the paragraph starting at line 116, it is really informative.

I had another look at the manuscript. I tried running the code line by line, but there are still some issues:

- Below line 107, before calling `data("ExampleData")`, you should add the line `library(surtvep)`.
- The code below line 110 doesn't work. I believe the correct function name is `IC`, so the line should be `fit.ic <- IC(fit.penalize)`.
- When I run `plot(fit.ic$mAIC, ylim = c(-3,10))` (below line 115), I get an empty graph. I think the correct code should be `plot(fit.ic$model.mAIC, ylim = c(-3,10))`

Finally, in the section "Availability", you wrote "Stable releases of the `surtvep` package will be made available via the Comprehensive R Archive Network." I think you can switch this to the present tense (i.e. "stable releases are available"), since it's now officially on CRAN!
