### 2023-12-04 #5688 @turgeonmaxime
@osorensen I would appreciate your input for some of the points in the checklist:

- **License file**: It's common for R packages to specify a license in the `DESCRIPTION` file, instead of having a separate copy of the license as part of the repository. Could you advise on whether @LingfengLuo0510 needs to include a copy of GPL-3 in the repo?
- **Testing**: There are no automated tests in the repo, but there are many examples that can be run manually to test the functionality. For that reason I checked the box, but let me know if that's not enough.
- **Contribution**: There is information on how to report bugs and ask for support, so again I checked the box. But there is no information on guidelines for people who would like to contribute. Could you advise on whether this is a strict requirement?

Apart from that, and provided @LingfengLuo0510 addresses the remaining issues from my previous comment, I'm done with the review.
