### 2023-12-05 #5688 @osorensen
Thanks for your review @turgeonmaxime. Here are my answers to your questions, which I hope can guide you on how to proceed further @LingfengLuo0510.

**License file**: It should by included. @LingfengLuo0510, please add a license file, and then also add the name of the license file to `.Rbuildignore`, so CRAN doesn't complain.

**Testing**: For R packages, there should be automated tests. @LingfengLuo0510, I suggest checking out the [testthat package](https://testthat.r-lib.org/) to get started. In addition to running on CRAN, the tests should also be part of an automated workflow in the GitHub repository, e.g., using GitHub Actions. You can read more on this in Wickham and Bryan's [R packages book](https://r-pkgs.org/).

**Contribution**: This should be added.
