+++ DESCRIPTION
@@ -18,7 +18,7 @@ Description: Fit Cox non-proportional hazards models with time-varying coefficie
     interval and hypothesis testing are also provided.
     For more information, see Wu et al. (2022) <doi: 10.1007/s10985-021-09544-2> and 
     Luo et al. (2023) <doi:10.1177/09622802231181471>.
-License: GPL-3
+License: GPL-3 | file LICENSE
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
