+++ vignettes/surtvep.Rmd
@@ -511,217 +511,217 @@ For the following plot, the knot was set as default, which is `knot=8`.
 
 
 
-<!-- ## Real Data Example `SUPPORT` -->
+## Real Data Example `SUPPORT`
 
-<!-- In this section, we illustrate the usage of our package by analyzing the time-varying effects on the real data `SUPPORT` (Study to Understand Prognoses Preferences Outcomes and Risks of Treatment). -->
-<!-- We use the processed `SUPPORT` data set included in the R package [@casebase]. -->
+In this section, we illustrate the usage of our package by analyzing the time-varying effects on the real data `SUPPORT` (Study to Understand Prognoses Preferences Outcomes and Risks of Treatment).
+We use the processed `SUPPORT` data set included in the R package [@casebase].
 
-<!-- We study the time-varying effects of metastatic cancer as the illustration example and include -->
+We study the time-varying effects of metastatic cancer as the illustration example and include
 
-<!-- First we load the data and get the survival outcome: -->
+First we load the data and get the survival outcome:
 
-<!-- ```{r, message = FALSE} -->
-<!-- data(support) -->
+```{r, message = FALSE}
+data(support)
 
-<!-- time <- support$d.time -->
-<!-- death <- support$death -->
-<!-- ``` -->
+time <- support$d.time
+death <- support$death
+```
 
-<!-- Age, cancer status and diabetes status are included as covariates and considered as categorical variables. -->
-<!-- Age has 4 levels, cancer status has 3 levels and diabetes has 2 levels. -->
-<!-- The following codes processed the covariates for model fitting. -->
+Age, cancer status and diabetes status are included as covariates and considered as categorical variables.
+Age has 4 levels, cancer status has 3 levels and diabetes has 2 levels.
+The following codes processed the covariates for model fitting.
 
-<!-- ```{r, message = FALSE} -->
-<!-- #diabetes: -->
-<!-- diabetes <-  model.matrix(~factor(support$diabetes))[,-1] -->
+```{r, message = FALSE}
+#diabetes:
+diabetes <-  model.matrix(~factor(support$diabetes))[,-1]
 
-<!-- #sex: female as the reference group -->
-<!-- sex <- model.matrix(~support$sex)[,-1] -->
+#sex: female as the reference group
+sex <- model.matrix(~support$sex)[,-1]
 
-<!-- #age: continuous variable -->
-<!-- age <-support$age -->
-<!-- age[support$age<=50] <- "<50" -->
-<!-- age[support$age>50 & support$age<=60] <- "50-59" -->
-<!-- age[support$age>60 & support$age<70] <- "60-69" -->
-<!-- age[support$age>=70] <- "70+" -->
-<!-- age <- factor(age, levels = c("60-69", "<50", "50-59", "70+")) -->
-<!-- z_age <- model.matrix(~age)[,-1] -->
+#age: continuous variable
+age <-support$age
+age[support$age<=50] <- "<50"
+age[support$age>50 & support$age<=60] <- "50-59"
+age[support$age>60 & support$age<70] <- "60-69"
+age[support$age>=70] <- "70+"
+age <- factor(age, levels = c("60-69", "<50", "50-59", "70+"))
+z_age <- model.matrix(~age)[,-1]
 
-<!-- # cancer status: metastatic, yes, no -->
-<!-- ca <- factor(support$ca) -->
-<!-- z_ca_mets <- as.numeric(support$ca == "metastatic") -->
-<!-- z_ca_non_mets <- as.numeric(support$ca == "yes") -->
+# cancer status: metastatic, yes, no
+ca <- factor(support$ca)
+z_ca_mets <- as.numeric(support$ca == "metastatic")
+z_ca_non_mets <- as.numeric(support$ca == "yes")
 
[…]
