+++ .github/workflows/R-CMD-check.yaml
@@ -1,0 +1,24 @@
+name: R-CMD-check
+
+on:
+  push:
+    branches:
+      - main
+      - CRAN
+  pull_request:
+
+jobs:
+  check:
+    runs-on: ubuntu-latest
+    container: 
+      image: rocker/r-ver:4.0.3
+    steps:
+      - uses: actions/checkout@v2
+      - name: Install dependencies
+        run: |
+          install2.r --error --deps TRUE testthat
+          install2.r --error --deps TRUE devtools
+      - name: Check package
+        run: |
+          R CMD build .
+          R CMD check *tar.gz
