+++ .github/workflows/R-CMD-check.yaml
@@ -11,7 +11,7 @@ jobs:
   check:
     runs-on: ubuntu-latest
     container: 
-      image: rocker/r-ver:4.0.3
+      image: rocker/r-ver:4.3.1
     steps:
       - uses: actions/checkout@v2
       - name: Install dependencies
