### 2024-02-16 #5688 @LingfengLuo0510
@turgeonmaxime @adibender @osorensen

In response to the review, we've incorporated testthat into our package and automated testing within our GitHub workflow. This ensures ongoing quality and facilitates future updates. We appreciate the feedback and are committed to continuous improvement. We will continue to update it in the future when adding new features. It is done for now for this paper review. Thanks a lot!
