### 2024-02-19 #5688 @adibender
> @adibender, could you please update [your checklist](https://github.com/openjournals/joss-reviews/issues/5688#issuecomment-1649418002) based on the latest changes made by the authors?

Hi, yes, will do sometime this week.
