+++ vignettes/surtvep.Rmd
@@ -1,0 +1,737 @@
+---
+title: "surtvep"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{surtvep}
+  %\VignetteEncoding{UTF-8}
+  %\VignetteEngine{knitr::rmarkdown}
+editor_options: 
+  markdown: 
+    wrap: sentence
+bibliography: references.bib
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+## Introduction:
+
+`surtvep` is an R package for fitting penalized Newton's method for the time-varying effects model using mAIC, TIC, GIC as information criteria, in particular we span the parameter using basis functions.
+Utilities for carrying out post-fitting visualization, summarization, and inference are also provided.
+In this tutorial we introduce the use of `surtvep` through an example dataset.
+
+This tutorial aims to provide a comprehensive guide on survival estimation for the time-varying coefficient model using the `surtvep` package in R. The package addresses the challenges associated with large-scale time-to-event data and the estimation of time-varying effects in survival analysis. By implementing a computationally efficient Kronecker product-based proximal algorithm and incorporating various penalties to improve estimation, `surtvep` offers an efficient and flexible solution for handling time-varying coefficients in survival analysis.
+
+Throughout the tutorial, we will cover the key features and capabilities of the `surtvep` package, providing examples and use cases to illustrate the package's functionality. By the end of this tutorial, you will have a solid understanding of how to use `surtvep` for survival estimation with time-varying coefficients, as well as how to leverage the package's unique features to improve your analysis.
+
+
+
+## Installation:
+
+```{r ,include=TRUE,eval=FALSE}
+#Install the package, need to install the devtools packages:
+require("devtools")
+require("remotes")
+remotes::install_github("UM-KevinHe/surtvep", ref = "Lingfeng_test")
+```
+
+## Quick Start
+
+The purpose of this section is to give users a general sense of the package.
+We will briefly go over the main functions, basic operations and outputs.
+After this section, users may have a better idea of what functions are available, which ones to use, or at least where to seek help.
+
+First, we load the 'surtvep' package:
+
+```{r}
+library("surtvep")
+```
+
+The main functions used in the package are Newton's method 'coxtv' and Newton's method combined with penalization 'coxtp', which we will demonstrate in this section.
+We load a set of data created beforehand for illustration:
+
+```{r}
+data("ExampleData")
+z     <- ExampleData$x
+time  <- ExampleData$time
+event <- ExampleData$event
+```
+
+The command loads an input covariate matrix 'z', time-to-event outcome 'time' and 'event' from this saved R data archive.
+The saved data set is a simulation data set with continuous covariates.
+
+We fit the Newton's method without penalization use the most basic call to 'coxtv'.
+
+```{r}
+fit.tv <- coxtv(z = z, event = event, time=time)
+```
+
+'fit.tv' is an object of class 'coxtv' that contains all the relevant information of the fitted model for further use.
+We do not encourage users to extract the components directly.
+Instead, various methods are provided for the object such as `plot` and `test` that enable us to execute those tasks more elegantly.
+
[…]
