+++ vignettes/references.bib
@@ -1,0 +1,8 @@
+@article{casebase,
+	title = {casebase: Fitting Flexible Smooth-in-Time Hazards and Risk Functions via Logistic and Multinomial Regression},
+	author = {Bhatnagar, Sahir and Turgeon, Maxime and Islam, Jesse and Saarela, Olli and Hanley, James},
+	year = {2020},
+	date = {2020},
+	url = {https://CRAN.R-project.org/package=casebase}
+}
+

[…]
