### 2024-03-08 ee68c0d
Fix some minor issues on 'Get started' page
