### 2024-03-08 08a4716
Fix some minor issues on 'Get started' page.
