### 2024-04-09 #5688 @adibender
@LingfengLuo0510 thanks for the updates.
Here are some additional notes.
There are still some points I haven't checked, because of the issues below, mostly concerning the paper and documentation.
Those concerned points are:
- reproducibility -> see comment regarding real data example
- performance (claims) -> see comments below
- state of the filed -> not really describe what alternative approaches exist
- references -> related to point above

The main improvement for the paper I think would be to make the the scope and potential limitations of the package more clear:
- you mention \beta(t) in the paper, but I think the types of models that can be fit should be made more clear, i.e. varying coefficient models of the form f(t)*x, i.e. (potentially) non-linear in time, but linear in x, in contrast for example to f(x)*t or f(t,x).
- Implementation wise, users can not choose which type of effect to fit, i.e. one can only specify covariates and they will automatically be assumed to have f(t)*x, but one cannot chose some x to be beta *x or interactions between different Xs, non-linear effects f(x), etc., which is fine, but should be briefly made explicit
- Line 44: "compared to existing computational packages..." Here are some performance claims, but the paper does not explicitly mention which packages were compared and I couldn't find the comparisons also in the referenced papers (see also my attached R code where I compare surtvep to 2 approaches, my own PAMMs and Aalens additive hazards model from timereg package, where for the simple case they appear to be faster)
- It's also not clear from the paper in which situations the package applies, n>p or n<<p?
- The Data Example is not reproducible as it only links to the Homepage, but not clear which data exactly was downloaded, how it was preprocessed, etc.
- It's not clear for which survival tasks it is applicable, only to right-censored data or also left-/interval censored, competing risks, multi-state models, etc.

To be clear, I don't ask to run some addtional models or change the implementation, but scope, limitations, overview of existing methods and performance claims should be adapted/verified in the paper.
@osorensen has to decide regarding the reproducibility aspect of the real data example.

# Some minor comments
- L59: "Breslow ... number of times" probably should be "number of ties"?
- L95: mentions some hypothesis testing capabilities, etc. (her references to the papers where these have been established)
- 104: Link in 104 leads to main page, therin is a section to "Detailed tutorial" but link goes back to landing page
- Figure 3 not reproducible: https://github.com/UM-KevinHe/surtvep/blob/JOSS/JOSS/GenerateRealDataPlot.R reads data from non-accessible location.
- for the figure here: https://um-kevinhe.github.io/surtvep/articles/surtvep.html#quick-start better to use `coord_cartesian(ylim = c(y1,y2))` rather than `ylim(c(y1,y2))`
