### 2024-05-08 #5688 @LingfengLuo0510
@adibender

1.	you mention \beta(t) in the paper, but I think the types of models that can be fit should be made more clear, i.e. varying coefficient models of the form f(t)*x, i.e. (potentially) non-linear in time, but linear in x, in contrast for example to f(x)*t or f(t,x).

Thank you for your suggestion. We have added a sentence at line 70 clarifying the type of models considered in this work.

2.	Implementation wise, users can not choose which type of effect to fit, i.e. one can only specify covariates and they will automatically be assumed to have f(t)*x, but one cannot chose some x to be beta *x or interactions between different Xs, non-linear effects f(x), etc., which is fine, but should be briefly made explicit

We have added a sentence at line 129 clarifying that the limitation of current implementation.

3.	Line 44: "compared to existing computational packages..." Here are some performance claims, but the paper does not explicitly mention which packages were compared and I couldn't find the comparisons also in the referenced papers (see also my attached R code where I compare surtvep to 2 approaches, my own PAMMs and Aalens additive hazards model from timereg package, where for the simple case they appear to be faster)

Thank you for bringing this up and for sharing your comparison code. We've clarified at line 45 that our performance comparisons were specifically against other packages for Cox proportional hazards models.

4.	It's also not clear from the paper in which situations the package applies, n>p or n<<p?

We have added relevant discussion at line 130 clarifying that the current implementation is for low-dimensional settings (where the number of covariates is much smaller than the sample size).

5.	The Data Example is not reproducible as it only links to the Homepage, but not clear which data exactly was downloaded, how it was preprocessed, etc.

The code for model fitting and plotting can be found in the Repository-joss branch. The name of the file is "GenerateRealDataPlot.R". Access to data can be requested at https://seer.cancer.gov/data/access.html. We have added a sentence at line 139 clarifying the data request. We also include the public available SUPPORT data example in the tutorial website.

6.	It's not clear for which survival tasks it is applicable, only to right-censored data or also left-/interval censored, competing risks, multi-state models, etc.

We have added a sentence at line 131 clarifying that the limitation of current implementation.

7.	L59: "Breslow ... number of times" probably should be "number of ties"?

Thank you for pointing it out. We changed the “times” to “ties” at Line 61.

8.	L95: mentions some hypothesis testing capabilities, etc. (add references to the papers where these have been established).

We have added the references giving the details of testing at line 129.

9.	L104: Link in 104 leads to main page, therein is a section to "Detailed tutorial" but link goes back to landing page

Thank you for pointing it out. We have updated the link.

10.	for the figure here: https://um-kevinhe.github.io/surtvep/articles/surtvep.html#quick-start better to use coord_cartesian(ylim = c(y1,y2)) rather than ylim(c(y1,y2))

Thank you for pointing it out. We will update this plot function in future releases of R-CRAN.
