### 2024-06-12 #5688 @adibender
@osorensen @LingfengLuo0510
my comments have been addressed sufficiently.
Just one more minor thing: In Line 37: "With the rising need for modeling time-varying effects, researchers have developed methods to
handle the data (Gray, 1992, 1994; Hastie & Tibshirani, 1993; Zucker & Karr, 1990)."
Rather than "the data" it should say "such data"? And even that would be a bit imprecise, because the data is standard, its the effects that change, so maybe reformulate a bit.
Other than that, good to go and I don't need to see another revision.

Sorry again for the delay.
