+++ JOSS/paper.md
@@ -41,7 +41,7 @@ The Cox non-proportional hazards model is a flexible and powerful tool for model
 
 Numerous studies have demonstrated the widespread presence of time-varying effects. For instance, the scientific literature has shown that factors like age, sex, and race can have non-constant associations with survival in cases such as end-stage renal disease [@he2017modeling; @he2022stratified], and breast cancer patients receiving neo-adjuvant chemotherapy and head and neck cancer patients [@baulies2015time; @brouwer2020time]. Ignoring the variations and relying solely on the Cox proportional hazards model can lead to inaccurate risk prediction and suboptimal treatment development.
 
-With the rising need for modeling time-varying effects, researchers have developed methods to handle the data [@hastie1993varying; @zucker1990nonparametric; @gray1992flexible; @gray1994spline]. In terms of implementation, these methods expand the original data in a repeated measurement format [@therneau2017using] using existing software such as the `survival` package [@survival-package]. Even with moderate sample sizes, this leads to a large and computationally burdensome working dataset. `surtvep` addresses this issue by implementing a computationally efficient Kronecker product-based proximal algorithm [@perperoglou2006fast], which can handle time-varying effects in large-scale studies with improved efficiency and parallel computing capabilities. Compared with existing computational packages for Cox non-proportional hazards models, such as the `coxph` function, `surtvep` demonstrates a much more efficient performance, with both runtime and memory consumption reduced considerably.
+With the rising need for modeling time-varying effects, researchers have developed methods to handle the complex and dynamic nature of such data [@hastie1993varying; @zucker1990nonparametric; @gray1992flexible; @gray1994spline]. In terms of implementation, these methods expand the original data in a repeated measurement format [@therneau2017using] using existing software such as the `survival` package [@survival-package]. Even with moderate sample sizes, this leads to a large and computationally burdensome working dataset. `surtvep` addresses this issue by implementing a computationally efficient Kronecker product-based proximal algorithm [@perperoglou2006fast], which can handle time-varying effects in large-scale studies with improved efficiency and parallel computing capabilities. Compared with existing computational packages for Cox non-proportional hazards models, such as the `coxph` function, `surtvep` demonstrates a much more efficient performance, with both runtime and memory consumption reduced considerably.
 
 Another issue of numerical instability arises when analyzing data with binary covariates that have limited variation. `surtvep` implements a proximal Newton's method to improve the estimation. Additionally, adding a penalty can improve the estimation. `surtvep` also supports P-spline and smoothing spline [@eilers1996flexible; @wood2017generalized; @wood2017p], to further improve estimation stability. The improved estimation performance of `surtvep` is demonstrated in our recent studies [@wu2022scalable; @lfluo2023smmr].
 
