### 2024-06-16 #5688 @LingfengLuo0510
@adibender Thank you for your suggestion. We have modified the sentence at line 37 to make it more clear.
The modified sentence is "With the rising need for modeling time-varying effects, researchers have developed methods to handle the complex and dynamic nature of such data."

@osorensen I have updated the above sentence. Thanks a lot!
