### 2024-06-19 #5688 @turgeonmaxime
> Checking the BibTeX entries failed with the following error:
>
> ```
> Failed to parse BibTeX on value "title" (NAME) [#<BibTeX::Bibliography data=[8]>, "@", #<BibTeX::Entry >, "%"]
> ```

@LingfengLuo0510 I think the bib entry that leads to a parsing issue is this one here:
https://github.com/UM-KevinHe/surtvep/blob/0ae36ca6d7a58aaaeb7902378438ea58611653ba/JOSS/paper.bib#L99-L107

Markup for comments in bibtex files is not standardized as far as I know. If you replace "@article" by "%article", or simply remove the @ sign, the bot should be able to compile your manuscript.
