### 2024-06-20 #5688 @osorensen
@LingfengLuo0510, I have now read through the manuscript once again, and think it's very well written. Below are some minor issues. Could you please address them, in addition to the [BibTeX issue](https://github.com/openjournals/joss-reviews/issues/5688#issuecomment-2179305784) mentioned above, and report here when done?

- I cannot see that Figure 1 or Figure 2 are mentioned anywhere in the text. Please add one sentence pointing the reader to each of them. For example "Figure 1 shows a flow chart for functions in the surtvep package."
- I suggest adding a reference to Eilers and Marx the first time you mention P-splines. Here is the original reference: https://doi.org/10.1214/ss/1038425655.
- On line 103, rewrite "covariates effects" to "covariate effects".
