+++ JOSS/paper.bib
@@ -1,34 +1,3 @@
-
-@article{gaia,
-    author = {{Gaia Collaboration}},
-    title = "{The Gaia mission}",
-    journal = {Astronomy and Astrophysics},
-    archivePrefix = "arXiv",
-    eprint = {1609.04153},
-    primaryClass = "astro-ph.IM",
-    keywords = {space vehicles: instruments, Galaxy: structure, astrometry, parallaxes, proper motions, telescopes},
-    year = 2016,
-    month = nov,
-    volume = 595,
-    doi = {10.1051/0004-6361/201629272},
-    url = {http://adsabs.harvard.edu/abs/2016A%26A...595A...1G},
-}
-
-@article{astropy,
-    author = {{Astropy Collaboration}},
-    title = "{Astropy: A community Python package for astronomy}",
-    journal = {Astronomy and Astrophysics},
-    archivePrefix = "arXiv",
-    eprint = {1307.6212},
-    primaryClass = "astro-ph.IM",
-    keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
-    year = 2013,
-    month = oct,
-    volume = 558,
-    doi = {10.1051/0004-6361/201322068},
-    url = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
-}
-
 @misc{fidgit,
   author = {A. M. Smith and K. Thaney and M. Hahnel},
   title = {Fidgit: An ungodly union of GitHub and Figshare},
@@ -96,15 +65,6 @@
   label={gray1992flexible}
 }
 
-% @article{gray1994spline,
-%   title={Spline-based tests in survival analysis},
-%   author={Gray, Robert J},
-%   journal={Biometrics},
-%   pages={640--652},
-%   year={1994},
-%   publisher={JSTOR},
-%   label={gray1994spline}
-% }
 
 @online{therneau2017using,
   author = {Therneau, Terry and Crowson, Cindy and Atkinson, Elizabeth},
