+++ JOSS/paper.bib
@@ -1,3 +1,18 @@
+@article{astropy,
+    author = {{Astropy Collaboration}},
+    title = "{Astropy: A community Python package for astronomy}",
+    journal = {Astronomy and Astrophysics},
+    archivePrefix = "arXiv",
+    eprint = {1307.6212},
+    primaryClass = "astro-ph.IM",
+    keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
+    year = 2013,
+    month = oct,
+    volume = 558,
+    doi = {10.1051/0004-6361/201322068},
+    url = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
+}
+
 @misc{fidgit,
   author = {A. M. Smith and K. Thaney and M. Hahnel},
   title = {Fidgit: An ungodly union of GitHub and Figshare},
@@ -16,7 +31,9 @@
   number={1},
   pages={30--36},
   year={2015},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi = {10.1051/0004-6361/201322068},
+  url = {https://doi.org/10.1038/bjc.2015.174}
 }
 
 
@@ -28,7 +45,9 @@
   number={23},
   pages={5137--5146},
   year={2020},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1002/cncr.33174},
+  url = {https://doi.org/10.1002/cncr.33174}
 }
 
 @article{zucker1990nonparametric,
@@ -39,7 +58,9 @@
   number={1},
   pages={329--353},
   year={1990},
-  publisher={Institute of Mathematical Statistics}
+  publisher={Institute of Mathematical Statistics},
+  doi = {10.1214/aos/1176347503},
+  url = {https://doi.org/10.1214/aos/1176347503}
 }
 
 @article{hastie1993varying,
@@ -50,7 +71,9 @@
   number={4},
   pages={757--779},
   year={1993},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1111/j.2517-6161.1993.tb01939.x},
+  url = {https://doi.org/10.1111/j.2517-6161.1993.tb01939.x}
 }
 
 @article{gray1992flexible,
@@ -62,7 +85,9 @@
   pages={942--951},
   year={1992},
   publisher={Taylor \& Francis},
-  label={gray1992flexible}
+  label={gray1992flexible},
+  doi = {10.2307/2290630},
+  url = {https://doi.org/10.2307/2290630}
 }
 
 
@@ -92,7 +117,9 @@
     number={2},
     pages={194--218},
     year={2022},
-    publisher={Springer}
+    publisher={Springer},
+    doi = {10.1007/s10985-021-09544-2},
+    url = {https://doi.org/10.1007/s10985-021-09544-2}
 }
 
 @book{wood2017generalized,
@@ -109,7 +136,9 @@
   volume={27},
   pages={985--989},
   year={2017},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s11222-016-9666-x},
+  url = {https://doi.org/10.1007/s11222-016-9666-x}
 }
 
 @misc{SEER,
@@ -126,7 +155,9 @@
   booktitle={{Selected Papers of Hirotogu Akaike}},
   pages={199--213},
   year={1998},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/978-1-4612-0919-5_38},
+  url = {https://doi.org/10.1007/978-1-4612-0919-5_38}
 }
 
 @article{eilers1996flexible,
@@ -137,7 +168,9 @@
[…]
