+++ JOSS/paper.bib
@@ -1,28 +1,3 @@
-@article{astropy,
-    author = {{Astropy Collaboration}},
-    title = "{Astropy: A community Python package for astronomy}",
-    journal = {Astronomy and Astrophysics},
-    archivePrefix = "arXiv",
-    eprint = {1307.6212},
-    primaryClass = "astro-ph.IM",
-    keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools},
-    year = 2013,
-    month = oct,
-    volume = 558,
-    doi = {10.1051/0004-6361/201322068},
-    url = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
-}
-
-@misc{fidgit,
-  author = {A. M. Smith and K. Thaney and M. Hahnel},
-  title = {Fidgit: An ungodly union of GitHub and Figshare},
-  year = {2020},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/arfon/fidgit}
-}
-
-
 @article{baulies2015time,
   title={Time-varying effect and long-term survival analysis in breast cancer patients treated with neoadjuvant chemotherapy},
   author={Baulies, S and Belin, L and Mallon, P and Senechal, C and Pierga, JY and Cottu, P and Sablin, MP and Sastre, X and Asselain, B and Rouzier, R and others},
