### 2024-06-21 #5688 @LingfengLuo0510
@osorensen

1. "I cannot see that Figure 1 or Figure 2 are mentioned anywhere in the text. Please add one sentence pointing the reader to each of them. For example "Figure 1 shows a flow chart for functions in the surtvep package.""

We have added sentences directing the reader to Figure 1 (line 62) and Figure 2 (line 121) within the relevant sections of the text.

2. "I suggest adding a reference to Eilers and Marx the first time you mention P-splines. Here is the original reference: https://doi.org/10.1214/ss/1038425655."

We have included the suggested reference to Eilers and Marx (1996) at the first mention of P-splines (line 18).

3. "On line 103, rewrite "covariates effects" to "covariate effects".

We have corrected the phrasing to "covariate effects" (line 103).

We have updated the .bib file as best as possible. Some references, such as those to R packages, do not have DOIs as they are not traditional journal articles. Is this ok?

Thanks a lot!
