### 2024-06-28 #5688 @LingfengLuo0510
> Yes please do so. I am getting ready to board a flight but will reaccept this as soon as I can. Just let me know once you have made the change.

@crvernon Than
Thanks a lot! I just updated the paper.md file
(1) Jeremy Taylor -> Jeremy M.G. Taylor
(2) Michael Kleinsasser -> Michael J. Kleinsasser

However, I'm encountering errors when generating the PDF, even after reverting to the previously accepted version. Could this be related to the paper's accepted status?
